[level unrated] Title: 옷가게 할인 받기, Time: 0.01 ms, Memory: 72.6 MB -BaekjoonHub

+++ 프로그래머스/unrated/120818. 옷가게 할인 받기/옷가게 할인 받기.java
@@ -1,7 +1,7 @@
 class Solution {
     public int solution(int price) {
         
-        if(price >= 500000) price = (int)price * 80/100;
+        if(price >= 500000) price = price * 80/100;
         
         else if(price >= 300000 && price < 500000) price = price * 90/100;
